Move ResourceGroupInfo to SPI
