+++ presto-main/src/main/java/com/facebook/presto/execution/resourceGroups/NoOpResourceGroupManager.java
@@ -16,6 +16,7 @@ package com.facebook.presto.execution.resourceGroups;
 import com.facebook.presto.execution.QueryExecution;
 import com.facebook.presto.spi.resourceGroups.ResourceGroupConfigurationManagerFactory;
 import com.facebook.presto.spi.resourceGroups.ResourceGroupId;
+import com.facebook.presto.spi.resourceGroups.ResourceGroupInfo;
 import com.facebook.presto.sql.tree.Statement;
 
 import java.util.concurrent.Executor;
